test suite for quadstore implementations with code generation
The qstest and testrunner packages provide a framework for running
generic tests on graph.QuadStore implementations.  And the accompanying
qstest command generates tests for implementations.

The command is compatible with `go generate` and generates a formatted
source file when given a testrunner.TestRunner variable name.

    //go:generate -runner=Runner -output quadstore_qstest_test.go

Eventually a default test runner will be built directly from a
qstest.QuadStoreImpl.